removed old test
