+++ python/cudf/cudf/tests/test_subword_tokenizer.py
@@ -1,12 +1,14 @@
 # Copyright (c) 2020-2021, NVIDIA CORPORATION.
 import os
 
+import cupy
 import numpy as np
 import pytest
 from transformers import BertTokenizer
 
 import cudf
 from cudf.core.subword_tokenizer import SubwordTokenizer
+from cudf.testing._utils import assert_eq
 
 
 @pytest.fixture(scope="module")
@@ -23,30 +25,6 @@ def assert_equal_tokenization_outputs(hf_output, cudf_output):
             hf_output["attention_mask"] != cudf_output["attention_mask"].get()
         )
         == 0
-    )
-
-
-def test_subword_tokenize_on_disk_vocab_str_api(datadir):
-    """
-    Tests the subword-tokenizer API where
-    the vocabulary is not pre-loaded
-    and is accessed via the string accessor
-    """
-    with open(
-        os.path.join(datadir, "test_sentences.txt"), encoding="utf-8"
-    ) as file:
-        input_sentence_ls = [line.strip() for line in file]
-
-    vocab_dir = os.path.join(datadir, "bert_base_cased_sampled")
-    vocab_hash_path = os.path.join(vocab_dir, "vocab-hash.txt")
-
-    ser = cudf.Series(input_sentence_ls)
-    tokens, masks, metadata = ser.str.subword_tokenize(
-        vocab_hash_path,
-        max_length=32,
-        stride=32,
-        do_lower=True,
-        max_rows_tensor=len(ser),
     )
 
 
@@ -115,3 +93,135 @@ def test_subword_tokenize_with_truncation(datadir):
             truncation=False,
             add_special_tokens=True,
         )
+
+
+def test_text_subword_tokenize(tmpdir):
+    sr = cudf.Series(
+        [
+            "This is a test",
+            "A test this is",
+            "Is test a this",
+            "Test   test",
+            "this   This",
+        ]
+    )
+    hash_file = tmpdir.mkdir("nvtext").join("tmp_hashed_vocab.txt")
+    content = "1\n0\n23\n"
+    coefficients = [65559] * 23
+    for c in coefficients:
+        content = content + str(c) + " 0\n"
+    # based on values from the bert_hash_table.txt file for the
+    # test words used here: 'this' 'is' 'a' test'
+    table = [0] * 23
+    table[0] = 3015668
+    table[1] = 6205475701751155871
+    table[5] = 6358029
+    table[16] = 451412625363
+    table[20] = 6206321707968235495
+    content = content + "23\n"
+    for v in table:
+        content = content + str(v) + "\n"
+    content = content + "100\n101\n102\n\n"
+    hash_file.write(content)
+
+    cudf_tokenizer = SubwordTokenizer("voc_hash.txt")
+
+    tokens, masks, metadata = cudf_tokenizer(sr, 8, 8)
+    expected_tokens = cupy.asarray(
+        [
+            2023,
+            2003,
+            1037,
+            3231,
+            0,
+            0,
+            0,
+            0,
+            1037,
+            3231,
+            2023,
+            2003,
+            0,
+            0,
+            0,
+            0,
+            2003,
+            3231,
+            1037,
+            2023,
+            0,
+            0,
+            0,
+            0,
+            3231,
+            3231,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            2023,
+            2023,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+        ],
+        dtype=np.uint32,
+    )
+    assert_eq(expected_tokens, tokens)
+
+    expected_masks = cupy.asarray(
+        [
+            1,
+            1,
+            1,
+            1,
+            0,
+            0,
+            0,
+            0,
+            1,
+            1,
+            1,
+            1,
+            0,
+            0,
+            0,
+            0,
+            1,
+            1,
+            1,
+            1,
+            0,
+            0,
+            0,
+            0,
+            1,
+            1,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            1,
+            1,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+        ],
+        dtype=np.uint32,
+    )
+    assert_eq(expected_masks, masks)
+
+    expected_metadata = cupy.asarray(
+        [0, 0, 3, 1, 0, 3, 2, 0, 3, 3, 0, 1, 4, 0, 1], dtype=np.uint32
+    )
+    assert_eq(expected_metadata, metadata)
